airflow/requirements.txt
pandas==2.2.3
pyarrow==18.1.0
psycopg2-binary==2.9.10
apache-airflow-providers-postgres
pyspark==3.5.3
SQLAlchemy>=1.4,<3.0